refactor: moving the media client
